requirements.txt
aioboto3==12.3.0
aiobotocore==2.11.2
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs==23.2.0
boto3==1.34.34
botocore==1.34.34
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm @ file:///croot/comm_1709322850197/work
contourpy==1.2.0
cycler==0.12.1
debugpy @ file:///work/perseverance-python-buildout/croot/debugpy_1698884710808/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deeplake==3.8.25
dill==0.3.8
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.13.1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.2.0
h5py==3.10.0
humbug==0.3.2
idna==3.6
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1704833016303/work
jedi @ file:///work/perseverance-python-buildout/croot/jedi_1698864582475/work
Jinja2==3.1.3
jmespath==1.0.1
jupyter_client @ file:///work/perseverance-python-buildout/croot/jupyter_client_1701731788627/work
jupyter_core @ file:///work/perseverance-python-buildout/croot/jupyter_core_1701731747496/work
kiwisolver==1.4.5
lz4==4.3.3
MarkupSafe==2.1.5
mat73==0.62
matplotlib==3.8.3
matplotlib-inline @ file:///work/perseverance-python-buildout/croot/matplotlib-inline_1698864771271/work
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.2.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1707225359967/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=031b7d6b2e5e604d9e21fc21be713ebf28ce133ec872dce6de006742d5e49bab
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0
opencv-python-headless==4.9.0
packaging==24.0
pandas==2.2.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathos==0.3.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==10.2.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pox==0.3.4
ppft==1.7.6.8
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
psutil @ file:///work/perseverance-python-buildout/croot/psutil_1698863411559/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pydantic==2.6.4
pydantic_core==2.16.3
Pygments @ file:///work/perseverance-python-buildout/croot/pygments_1698846270603/work
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
requests==2.31.0
s3transfer==0.10.1
scipy==1.12.0
seaborn==0.13.2
setuptools==68.2.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.12
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tornado @ file:///work/perseverance-python-buildout/croot/tornado_1698866362018/work
tqdm==4.66.2
traitlets @ file:///work/perseverance-python-buildout/croot/traitlets_1698846387632/work
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
wheel==0.41.2
wrapt==1.16.0
yarl==1.9.4
